+++ lib/logstorage/parser.go
@@ -786,6 +786,38 @@ func ParseStatsQuery(s string) (*Query, error) {
 	return q, nil
 }
 
+// ContainAnyTimeFilter returns true when query contains a global time filter.
+func (q *Query) ContainAnyTimeFilter() bool {
+	if hasTimeFilter(q.f) {
+		return true
+	}
+	for _, p := range q.pipes {
+		if pf, ok := p.(*pipeFilter); ok {
+			if hasTimeFilter(pf.f) {
+				return true
+			}
+		}
+	}
+	return false
+}
+
+func hasTimeFilter(f filter) bool {
+	if f == nil {
+		return false
+	}
+	switch t := f.(type) {
+	case *filterAnd:
+		for _, subF := range t.filters {
+			if hasTimeFilter(subF) {
+				return true
+			}
+		}
+	case *filterTime:
+		return true
+	}
+	return false
+}
+
 // ParseQueryAtTimestamp parses s in the context of the given timestamp.
 //
 // E.g. _time:duration filters are adjusted according to the provided timestamp as _time:[timestamp-duration, duration].